Fade animation renamed and float shader property animation implemented.
