+++ Scripts/FloatShaderPropertyAnimation.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d3597b1aef2e2c34d84db29f91960dcd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
